PTG Iterator

+++ torch_geometric_temporal/data/splitter/discrete_splitter.py
@@ -1,6 +1,6 @@
 from torch_geometric_temporal.data.discrete.static_graph_discrete_signal import StaticGraphDiscreteSignal
 
 
-def discrete_train_test_split(discrete_iterator, train_ratio: int=0.8):
+def discrete_train_test_split(discrete_iterator, train_ratio: float=0.8):
     print(type(discrete_iterator))
     return 1
